Build the project using Maven
This introduces a pom.xml which tells Maven how the
project is structured. To avoid a boatload of disruptive
renames, we override some of Maven's default settings.

It updates the code generation scheme to generate
VersionInfo.java using the templating-maven-plugin.

Lastly, it removes the Ant build files to avoid
maintaining multiple build systems.